Updated Unit test for refactoring.
- Moved test methods to their correct test classes.
- Created/Implemented Unit test for new TeacherService.
Refactor map:
TeacherServiceTest -> StudentServiceTest
StudentServiceTest -> AssignmentServiceTest
